requirements.txt
# AgentCert MVP Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2
openai==1.3.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dateutil==2.8.2

